Update README.md
he E-Learning Platform is a dynamic and scalable online education system designed to make learning accessible, engaging, and effective. It is a full-stack web application that connects students, educators, and administrators in a seamless digital learning ecosystem. The platform empowers users with tools to create, manage, and consume educational content across diverse fields of study.